+++ src/lib/lib.ts
@@ -1,4 +1,4 @@
-import { TLLCoordinates, TMapInfo, TXYCoordinates } from "./types/locationTypes";
+import { TLLCoordinates, TMapInfo, TXYCoordinates } from "../types/locationTypes";
 
 export const deg2rad = (deg: number) => deg * (Math.PI / 180);
 
@@ -34,25 +34,3 @@ export const getDistanceFromLatLonInM = (
   const d = earthRadius * c * 1000; // Distance in m
   return d;
 };
-
-// export function getDistance(
-//   { lat: lat1, lon: lon1 }: TLLCoordinates,
-//   { lat: lat2, lon: lon2 }: TLLCoordinates
-// ) {
-//   const radLat1 = (Math.PI * lat1) / 180;
-//   const radLat2 = (Math.PI * lat2) / 180;
-//   const theta = lon1 - lon2;
-//   const radTheta = (Math.PI * theta) / 180;
-//   let dist =
-//     Math.sin(radLat1) * Math.sin(radLat2) +
-//     Math.cos(radLat1) * Math.cos(radLat2) * Math.cos(radTheta);
-//   if (dist > 1) dist = 1;
-
-//   dist = Math.acos(dist);
-//   dist = (dist * 180) / Math.PI;
-//   dist = dist * 60 * 1.1515 * 1.609344 * 1000;
-//   // if (dist < 100) dist = Math.round(dist / 10) * 10;
-//   // else dist = Math.round(dist / 100) * 100;
-
-//   return dist;
-// }
